+++ plat/rockchip/rk3399/include/shared/pmu_regs.h
